+++ crates/rattler_solve/CHANGELOG.md
@@ -6,6 +6,36 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.6](https://github.com/conda/rattler/compare/rattler_solve-v1.4.5...rattler_solve-v1.4.6) - 2025-05-03
+
+### Other
+
+- lock workspace member dependencies ([#1279](https://github.com/conda/rattler/pull/1279))
+
+## [1.4.5](https://github.com/conda/rattler/compare/rattler_solve-v1.4.4...rattler_solve-v1.4.5) - 2025-04-17
+
+### Other
+
+- bump resolvo 0.9.0 ([#1259](https://github.com/conda/rattler/pull/1259))
+
+## [1.4.4](https://github.com/conda/rattler/compare/rattler_solve-v1.4.3...rattler_solve-v1.4.4) - 2025-04-10
+
+### Other
+
+- update Cargo.toml dependencies
+
+## [1.4.3](https://github.com/conda/rattler/compare/rattler_solve-v1.4.2...rattler_solve-v1.4.3) - 2025-04-04
+
+### Other
+
+- update Cargo.toml dependencies
+
+## [1.4.2](https://github.com/conda/rattler/compare/rattler_solve-v1.4.1...rattler_solve-v1.4.2) - 2025-03-14
+
+### Other
+
+- updated the following local packages: rattler_conda_types
+
 ## [1.4.1](https://github.com/conda/rattler/compare/rattler_solve-v1.4.0...rattler_solve-v1.4.1) - 2025-03-10
 
 ### Other